Cleaned Logo component & moved under new icons folder
