+++ src/components/atomic/icons/Logo.tsx
@@ -1,14 +1,21 @@
 interface LogoProps {
-    className?: string;
+  className?: string;
 }
 
 function Logo({ className }: LogoProps) {
-    return ( 
-        <svg className={className} version="1.0" xmlns="http://www.w3.org/2000/svg"
-     width="1500.000000pt" height="785.000000pt" viewBox="0 0 1500.000000 785.000000"
-     preserveAspectRatio="xMidYMid meet">
-        <g transform="translate(0.000000,785.000000) scale(0.100000,-0.100000)">
-            <path d="M3050 7846 c-3 -2 -187 -7 -410 -10 -223 -4 -454 -9 -515 -11 -60 -3
+  return (
+    <svg
+      className={className}
+      version="1.0"
+      xmlns="http://www.w3.org/2000/svg"
+      width="1500.000000pt"
+      height="785.000000pt"
+      viewBox="0 0 1500.000000 785.000000"
+      preserveAspectRatio="xMidYMid meet"
+    >
+      <g transform="translate(0.000000,785.000000) scale(0.100000,-0.100000)">
+        <path
+          d="M3050 7846 c-3 -2 -187 -7 -410 -10 -223 -4 -454 -9 -515 -11 -60 -3
             -283 -7 -495 -11 -212 -3 -419 -7 -460 -9 -41 -2 -273 -6 -515 -9 -242 -4
             -468 -9 -501 -12 -70 -6 -104 -29 -104 -71 0 -23 130 -257 518 -928 57 -99
             132 -229 166 -290 35 -60 107 -186 161 -280 53 -93 109 -190 123 -215 13 -25
@@ -91,16 +98,21 @@ function Logo({ className }: LogoProps) {
             299 11 430 74 199 306 396 519 440 73 15 303 4 407 -19z m903 -1618 c99 -26
             180 -50 180 -54 0 -6 -109 -50 -210 -84 -177 -60 -430 -76 -587 -37 -85 22
             -178 64 -227 105 l-29 24 49 12 c27 8 78 21 114 31 161 45 259 58 417 54 97
-            -2 136 -9 293 -51z"/>
-            <path d="M8235 1571 c-133 -33 -260 -188 -335 -411 -32 -94 -30 -132 9 -171
+            -2 136 -9 293 -51z"
+        />
+        <path
+          d="M8235 1571 c-133 -33 -260 -188 -335 -411 -32 -94 -30 -132 9 -171
             50 -49 114 -38 228 39 32 22 97 62 143 90 186 110 257 238 200 357 -41 84
-            -138 122 -245 96z"/>
-            <path d="M6325 1551 c-53 -24 -101 -75 -122 -129 -14 -38 -15 -53 -5 -98 19
+            -138 122 -245 96z"
+        />
+        <path
+          d="M6325 1551 c-53 -24 -101 -75 -122 -129 -14 -38 -15 -53 -5 -98 19
             -90 53 -130 192 -221 69 -45 145 -101 169 -123 57 -55 83 -64 127 -45 74 31
-            88 128 38 267 -39 109 -57 139 -141 239 -105 126 -167 152 -258 110z"/>
-        </g>
+            88 128 38 267 -39 109 -57 139 -141 239 -105 126 -167 152 -258 110z"
+        />
+      </g>
     </svg>
-     );
+  );
 }
 
 export default Logo;
